Reads and summarizes ProxiMate spectroscopic applications (nax files)

This function reads and summarizes the main aspects of BUCHI ProxiMate applications which are files of extension .nax. In addition, the file is retain as raw binary in the object generated by this function.

Usage

proximate_read_nax(file, ignore_version = FALSE)

Arguments

file

a character vector containing the .nax file name (and path).

ignore_version

a logical passed to proximate_read_cal. If FALSE (default), embedded .cal files with no version information or created with NIRWise PLUS prior to version 1.0 raise an error. Set to TRUE to read them with a warning instead.

Value

A a list of class nax which contains the following objects:

nax_summary: a list with:

content: the name of the files inside the nax/application.

size: the size (on disk) of the nax.

raw: the original nax file/application stored as raw binary.

nad_info: a list with:

summary: a summary of the high-level ProxiMate application parameters.

data: a full list of the high-level ProxiMate application parameters.

cal_info: a list with:

summary: a summary of the calibration models contained in the ProxiMate application.

meta_param: a list with parameters of each calibration model (e.g. pre-processing recipes.

rtf_info: a list with:

summary: a summary of the calibration models as printed in the calibration reports contained in the nax file. This includes the optimal number of components suggested (ncomp).

data: a list with:

summary: a summary of the calibration data (tsv files) contained in the ProxiMate application.

data: a list with the calibration data found in all the tsv files.

In case, any of the above components is encrypted a character string indicating so will be returned. In case of the rtf calibration reports are not present in the nax, a NULL will be returned for rtf_info.

Author

Leonardo Ramirez-Lopez