Recalibrate a nax file

This function updates a nax file

Usage

proximate_recalibrate_nax(x,
                          preprocess_recipes = NULL,
                          methods = NULL,
                          control = calibration_control(seed = 1),
                          name,
                          add = NULL)

Arguments

x

an object of class nax as returned by the proximate_read_nax function.

preprocess_recipes

an optional list with one or more objects of class preprocess_recipe that are to be tested for finding the optimal one for each model in the list passed to formulas.

methods

an optional list containing one ore more objects of class fit_constructor which are as returned by one of the fit_constructors functions, indicating what type of regression method to use along with its parameters.

control

a calibration_control object as returned by the calibration_control function, indicating how some aspects of the calibration process must be conducted (e.g. cross-validation and outlier detection). Default is calibration_control(seed = 1). See details.

name

a vector length at most 2, consisting of characters for the name and alias of the application. Defaults to "Untitled".

add

an optional object of class nax_augment as returned by the proximate_add2nax function.

Value

A list of class "spectral_multimodel". See calibrate_models function.

See also

Author

Leonardo Ramirez-Lopez and Claudio Orellano