Overview of the proximetricsR package

NIR calibration and application tools for BUCHI ProxiMate and ProxiScout devices.

Details

This is package version 0.7.1 (Valinhos).

This package provides R functions for spectral pre-processing, NIR model calibration, and reading/writing files for BUCHI ProxiMate and ProxiScout devices. The calibration algorithms (fit_plsr, fit_xlsr) and the pre-treatment constructors (prep_smooth, prep_snv, prep_resample, prep_derivative) reproduce the corresponding algorithms in BUCHI NIRWise PLUS (version 1.1.3000.0), guaranteeing numerical compatibility between models built with this package and those built in NIRWise PLUS.

The ProxiScout functions for preprocessing are also numerically equivalent to the ones of the "BUCHI Modeller" software. The regression method in te Modeller is teh classical PLS regression, however, the other PLS algorithms implemented in proximetricsR (modified PLS, standard PLS, and XLS) can also be used to generate models for ProxiScout devices.

The functions available for ProxiMate spectral data are:

proximate_read_data

proximate_data

proximate_merge

The functions available for reading generic spectral data files are:

read_spc

The functions available for spectral pre-processing are:

prep_resample

prep_smooth

prep_snv

prep_derivative

prep_detrend

prep_transform

prep_wav_trim

preprocess_recipe

process

The functions available for calibrating NIR regression models are:

calibrate

calibrate_models

calibration_control

fit_plsr

fit_xlsr

add_model_metadata

validate_prediction

The functions available for writing ProxiMate files are:

proximate_write_data

proximate_write_model

add_application_metadata

proximate_write_nax

The functions available for reading and editing ProxiMate application files are:

proximate_read_cal

proximate_read_nax

proximate_recalibrate_nax

proximate_add2nax

The functions available for ProxiScout devices are:

proxiscout_read_data

proxiscout_write_data

proxiscout_write_model

get_proxiscout_wavenumbers

proxiscout_repetition_pattern

The functions available for creating plots are:

plot.spectral_model

Other functions:

extract_property_names

A typical example dataset for a ProxiMate device can be found in:

proximateCannabis

A typical example dataset for a ProxiScout device can be found in:

proxiscoutCannabis

See also

Author

Leonardo Ramirez-Lopez, Claudio Orellano, Nicolae Cudlenco, Mai Said, Mohamed Abushosha, Marcal Plans