Calculate the ASCII Value Difference Between Two Strings

This function calculates the difference between two strings based on their ASCII values.

Usage

string_diff(s1, s2)

Arguments

s1

A character string.

s2

A character string.

Value

The absolute difference between the ASCII values of the characters in the two strings.

Details

The function ensures that both strings are of the same length by padding them with spaces if necessary. It then computes the difference between the ASCII values of corresponding characters in the strings.