Adds a chosen number of columns with entries equal to zero to a matrix object

Adds a chosen number of columns with entries equal to zero to a matrix object

Usage

add_zero_cols(object, n_zero_cols)

Arguments

object

matrix

n_zero_cols

an integer for the number of columns to be added

Value

the matrix with n_zero_cols zero-columns added to the left and right side of object